task 9
task 9

+++ server.js
@@ -2,6 +2,9 @@ const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
 app.use(bodyParser.json());
+let auth = require('./auth')(app);
+const passport = require('passport');
+require('./passport');
 
 const mongoose = require('mongoose');
 const Models = require('./models.js');
@@ -35,14 +38,32 @@ app.post('/users', async (req, res) => {
     }
 });
 
-// UPDATE
-// Update user info by user ID
-app.put('/users/:id', async (req, res) => {
+
+
+
+// Update user info by username
+app.put('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    // Check if the authenticated user is the same as the user to be updated
+    if (req.user.Username !== req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+
     try {
-        const userId = req.params.id;
-        const updatedUser = await Users.findByIdAndUpdate(userId, req.body, { new: true });
+        const updatedUser = await Users.findOneAndUpdate(
+            { Username: req.params.Username },
+            {
+                $set: {
+                    Username: req.body.Username,
+                    Password: req.body.Password,
+                    Email: req.body.Email,
+                    Birthday: req.body.Birthday
+                }
+            },
+            { new: true }
+        );
+
         if (updatedUser) {
-            res.status(200).json(updatedUser);
+            res.json(updatedUser);
         } else {
             res.status(404).send('No such user');
         }
@@ -53,13 +74,19 @@ app.put('/users/:id', async (req, res) => {
 });
 
 // ADD a movie to a user's list of favorites
-app.post('/users/:Username/movies/:MovieID', async (req, res) => {
+app.post('/users/:Username/movies/:MovieID', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    // Check if the authenticated user is the same as the user adding the favorite movie
+    if (req.user.Username !== req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+
     try {
         const updatedUser = await Users.findOneAndUpdate(
             { Username: req.params.Username },
-            { $push: { FavoriteMovies: new mongoose.Types.ObjectId(req.params.MovieID) } }, // Correct usage
+            { $push: { FavoriteMovies: new mongoose.Types.ObjectId(req.params.MovieID) } },
             { new: true }
         );
+
         if (updatedUser) {
             res.json(updatedUser);
         } else {
@@ -72,13 +99,19 @@ app.post('/users/:Username/movies/:MovieID', async (req, res) => {
 });
 
 // DELETE (Remove a movie from a user's favorite list)
-app.delete('/users/:Username/movies/:MovieID', async (req, res) => {
+app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    // Check if the authenticated user is the same as the user deleting the favorite movie
+    if (req.user.Username !== req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+
     try {
         const updatedUser = await Users.findOneAndUpdate(
             { Username: req.params.Username },
-            { $pull: { FavoriteMovies: new mongoose.Types.ObjectId(req.params.MovieID) } }, // Correct usage
+            { $pull: { FavoriteMovies: new mongoose.Types.ObjectId(req.params.MovieID) } },
             { new: true }
         );
+
         if (updatedUser) {
             res.json(updatedUser);
         } else {
@@ -91,7 +124,12 @@ app.delete('/users/:Username/movies/:MovieID', async (req, res) => {
 });
 
 // DELETE a user by username
-app.delete('/users/:Username', async (req, res) => {
+app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    // Check if the authenticated user is the same as the user being deleted
+    if (req.user.Username !== req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+
     try {
         const user = await Users.findOneAndDelete({ Username: req.params.Username });
         if (!user) {
@@ -107,18 +145,39 @@ app.delete('/users/:Username', async (req, res) => {
 
 // READ 
 // Get all movies
-app.get('/movies', async (req, res) => {
+app.get('/movies', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    await Movies.find()
+        .then((movies) => {
+            res.status(201).json(movies);
+        })
+        .catch((error) => {
+            console.error(error);
+            res.status(500).send('Error: ' + error);
+        });
+});
+
+// Get a user by username
+app.get('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    // Check if the authenticated user is the same as the requested user
+    if (req.user.Username !== req.params.Username) {
+        return res.status(400).send('Permission denied');
+    }
+
     try {
-        const movies = await Movies.find();
-        res.status(200).json(movies);
-    } catch (err) {
-        console.error(err);
-        res.status(500).send('Error: ' + err);
+        const user = await Users.findOne({ Username: req.params.Username });
+        if (user) {
+            res.json(user);
+        } else {
+            res.status(404).send('User not found');
+        }
+    } catch (error) {
+        console.error(error);
+        res.status(500).send('Error: ' + error);
     }
 });
 
 // Get a single movie by title
-app.get('/movies/:title', async (req, res) => {
+app.get('/movies/:title', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
         const movie = await Movies.findOne({ Title: req.params.title });
         if (movie) {
@@ -133,7 +192,7 @@ app.get('/movies/:title', async (req, res) => {
 });
 
 // Get genre data by name
-app.get('/movies/genre/:genreName', async (req, res) => {
+app.get('/movies/genre/:genreName', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
         const genre = await Genres.findOne({ Name: req.params.genreName });
         if (genre) {
@@ -148,7 +207,7 @@ app.get('/movies/genre/:genreName', async (req, res) => {
 });
 
 // Get director data by name
-app.get('/movies/directors/:directorName', async (req, res) => {
+app.get('/movies/directors/:directorName', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
         const director = await Director.findOne({ Name: req.params.directorName });
         if (director) {
@@ -161,5 +220,6 @@ app.get('/movies/directors/:directorName', async (req, res) => {
         res.status(500).send('Error: ' + err);
     }
 });
+
 
 app.listen(8080, () => console.log("Listening on port 8080"));
